requirements.txt
discord.py>=2.3.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
python-magic>=0.4.27
pypdf>=3.17.0
yt-dlp>=2024.1.0
# News DM subscriptions (RSS); auto-installed on startup if missing — see utils/bootstrap_deps.py
feedparser>=6.0.0
beautifulsoup4>=4.12.0
